Repurpose conf/extra directory for httpd config
Change the conf/extra directory from autoloading to holding files to be loaded
conditionally, or as needed.

Add conf/conf.d/config-extra.conf as a place to globally load files from the
repurposed conf/extra directory.

Change the order of the conf.d and extra directories in the httpd root.

Adjust the httpd.conf header to reflect the new order of things.

Extract the extended status flag from httpd-status.conf enabling the status to
be limited to a vhost while having extended status available.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>